Agrega y soluciona navegación para ver detalles de una meta
